--- client/components/Home.js
@@ -1,12 +1,0 @@
-import React, {Component} from 'react';
-
-
-export default class Home extends Component {
-  render () {
-    return (
-      <div className='page home'>
-        Home Page
-      </div>
-    );
-  }
-}
