+++ client/components/Navbar.js
@@ -1,7 +1,12 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-
+/**
+ * Navbar Component:
+ *   Renders the navbar
+ *
+ *   exported to /client/routes.js
+ */
 export default function Navbar () {
   return (
     <div className='navbar'>
