--- client/components/SignUp.js
@@ -1,12 +1,0 @@
-import React, {Component} from 'react';
-
-
-export default class SignUp extends Component {
-  render () {
-    return (
-      <div className='page'>
-        SignUp Page
-      </div>
-    );
-  }
-}
