+++ client/containers/index.js
@@ -1,0 +1,1 @@
+export {default as BuildBox} from './BuildBox';
